--- rapid_layout/default_models.yaml
@@ -1,4 +1,0 @@
-pp_layout_cdla:
-  model_path:
-  sha256:
-
