+++ rapid_layout/process/doc_layout/__init__.py
@@ -1,0 +1,5 @@
+# -*- encoding: utf-8 -*-
+# @Author: SWHL
+# @Contact: liekkaskono@163.com
+from .pose_process import DocLayoutPostProcess
+from .pre_process import DocLayoutPreProcess
