+++ rapid_layout/process/pp/pre_process.py
@@ -7,8 +7,6 @@ from typing import Optional, Tuple, Union
 import cv2
 import numpy as np
 
-from .augment import LetterBox
-
 InputType = Union[str, np.ndarray, bytes, Path]
 
 
@@ -39,30 +37,3 @@ class PPPreProcess:
 
     def permute(self, img: np.ndarray) -> np.ndarray:
         return img.transpose((2, 0, 1))
-
-
-class YOLOv8PreProcess:
-    def __init__(self, img_size: Tuple[int, int]):
-        self.img_size = img_size
-
-    def __call__(self, image: np.ndarray) -> np.ndarray:
-        input_img = cv2.resize(image, self.img_size)
-        input_img = input_img / 255.0
-        input_img = input_img.transpose(2, 0, 1)
-        input_tensor = input_img[np.newaxis, :, :, :].astype(np.float32)
-        return input_tensor
-
-
-class DocLayoutPreProcess:
-    def __init__(self, img_size: Tuple[int, int]):
-        self.img_size = img_size
-        self.letterbox = LetterBox(new_shape=img_size, auto=False, stride=32)
-
-    def __call__(self, image: np.ndarray) -> np.ndarray:
-        input_img = self.letterbox(image=image)
-        input_img = input_img[None, ...]
-        input_img = input_img[..., ::-1].transpose(0, 3, 1, 2)
-        input_img = np.ascontiguousarray(input_img)
-        input_img = input_img / 255
-        input_tensor = input_img.astype(np.float32)
-        return input_tensor
