+++ rapid_layout/process/yolov8/__init__.py
@@ -1,0 +1,5 @@
+# -*- encoding: utf-8 -*-
+# @Author: SWHL
+# @Contact: liekkaskono@163.com
+from .post_process import YOLOv8PostProcess
+from .pre_process import YOLOv8PreProcess
